pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = "A Streamlit app using Langchain, GPT-4, DuckDuckGo for real-time weather with memory."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
langchain = ">=0.3.24,<1.0.0"
openai = "^1.17.4"
duckduckgo-search = "^8.0.1"
tiktoken = ">=0.7,<1"
streamlit = "^1.45.0"
python-dotenv = "^1.1.0"
langchain-community = "^0.3.23"
langchain-openai = "^0.3.16"
google-api-python-client = "^2.169.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
black = "^22.3"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"